pyproject.toml
[project]
name = "pret-a-depenser"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
  "fastapi>=0.110,<1.0",
  "uvicorn[standard]>=0.20,<1.0",
  "pydantic>=2.0,<3.0",
  "python-dotenv>=1.0,<2.0",
  "huggingface_hub>=0.20,<1.0",
  "numpy>=1.26,<2.0",
  "pandas>=2.0,<3.0",
  "scikit-learn>=1.3,<1.6",
  "catboost>=1.2,<2.0",
  "psycopg[binary]>=3.1,<4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<10.0",
  "pytest-cov>=4.0,<8.0",
  "httpx>=0.24,<1.0",
]

notebooks = [
  "matplotlib>=3.7,<4.0",
  "seaborn>=0.13,<1.0",
  "mlflow>=2.9,<3.0",
  "shap>=0.44,<1.0",
  "xlsxwriter>=3.0,<4.0",
  "optuna",
  "xgboost>=2.0,<3.0",
  "lightgbm>=4.0,<5.0",
  "torch>=2.1,<3.0",
]

monitoring = [
  "streamlit>=1.30,<2.0",
  "matplotlib>=3.7,<4.0",
  "pandas>=2.0,<3.0",

]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "core*"]